define date time format consistently
Use consistent datetime formatting instead of hardcoded '+02:00', which might be understood as the requirement to always deliver a fixed time zone.

Always use upper case date and lower case time like in the referred wikipedia article. The timezone specification is also according to the wikipedia article.
Additionally fix the example date '2016-01-16T16:26:15+02:00' to be in June because the time zone offset doesn't make sense for Switzerland in January.

Use the name from wikipedia ('Date and time with the offset') for the format as well.